sms does not have src/morphology/stems/abbreviations.lexc

+++ corpus/analyse.sh
@@ -25,9 +25,12 @@ do
     # build fst needed for analysis
     cd $GTHOME/langs/$lang
     make
-    cd $GTHOME/langs/$lang/tools/preprocess
-    rm abbr.txt
-    make abbr
+    if [ $lang != "sms" ]
+    then
+        cd $GTHOME/langs/$lang/tools/preprocess
+        rm abbr.txt
+        make abbr
+    fi
 
 
     for corpus in $GTFREE $GTBOUND
